requirements.txt
# Core
python-dotenv>=1.0.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.3.0

# NLP & LLMs
transformers>=4.36.0
torch>=2.0.0
sentence-transformers>=2.2.2

# RAG & Orchestration
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0

# Vector Database
faiss-cpu>=1.7.4

# Document Processing
openpyxl>=3.1.0
pypdf>=3.17.0
tqdm>=4.65.0

# Evaluation
rouge>=1.0.1
nltk>=3.8.0
bert-score>=0.3.13
ragas>=0.1.0
datasets>=2.14.0

# Web UI
chainlit>=2.0.0

# Notebook Environment
jupyter>=1.0.0
ipykernel>=6.25.0
